% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transformations.R
\name{arcs_to_bearings}
\alias{arcs_to_bearings}
\title{Function to transform arcs to bearings}
\usage{
arcs_to_bearings(omega, delta, hbear0, vbear0)
}
\arguments{
\item{omega}{arc size at time t}

\item{delta}{arc orientation at time t}

\item{hbear0}{previous horizontal bearing (at time t-1)}

\item{vbear0}{previous vertical bearing (at time t-1)}
}
\description{
Function to transform arcs to bearings
}
