% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_dt}
\alias{calc_dt}
\title{Calculate difference in time}
\usage{
calc_dt(data, units = "hours", dir = "next")
}
\arguments{
\item{data}{data frame of locs with ID and time columns}

\item{units}{units of time (default = "hours")}

\item{dir}{calculate difference to next or previous loc? (default = "next")}
}
\description{
Calculate difference in time
}
