% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_betas.R
\name{correct_betas}
\alias{correct_betas}
\title{Correct estimated parameters for a gamma step length distribution 
and (potentially) a Kent arc distribution}
\usage{
correct_betas(model, data)
}
\arguments{
\item{model}{fitted model output}

\item{data}{data used to fit model}
}
\description{
Correct estimated parameters for a gamma step length distribution 
and (potentially) a Kent arc distribution
}
