% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dKent.R
\name{dKent}
\alias{dKent}
\title{Density of a Kent distribution}
\usage{
dKent(kappa, rho, omega, delta, log = FALSE)
}
\arguments{
\item{kappa}{angular concentration parameter}

\item{rho}{ovalness parameter}

\item{omega}{arc size}

\item{delta}{arc orientation}

\item{log}{whether to return the log density or not}
}
\value{
density or log density
}
\description{
Density of a Kent distribution
}
