% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_3d2.R
\name{extract_3d2}
\alias{extract_3d2}
\title{Extract 3D covariates from raster stack}
\usage{
extract_3d2(raster_stack, points)
}
\arguments{
\item{raster_stack}{SpatRaster stack; must have "depth", which is used
to find layer associated with a given z value}

\item{points}{Matrix or data frame with three columns, for x, y and z
coordinates}
}
\value{
Vector of covariate values at points.
}
\description{
Extract 3D covariates from raster stack
}
