% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transformations.R
\name{get_cartesian}
\alias{get_cartesian}
\title{Get Cartesian coordinates from spherical coordinates}
\usage{
get_cartesian(coord1, coord2, type = "geodesics")
}
\arguments{
\item{coord1}{first spherical coordinate}

\item{coord2}{second spherical coordinate}

\item{type}{either "bearings" or "geodesics"}
}
\value{
Matrix with three columns for Cartesian coordinates
}
\description{
Get Cartesian coordinates from spherical coordinates
}
\details{
The formulas depends on the 'type' argument. If type = "geodesics",
coord1 measures the "colatitude" angle between 0 and pi, and the 
transformations are:
\itemize{
\item cart1 = sin(coord1)cos(coord2)
\item cart2 = sin(coord1)sin(coord2)
\item cart3 = cos(coord1)
}
If type = "bearings", coord1 is the "latitude" angle between -pi/2 and pi/2,
and the transformations are: 
\itemize{
\item cart1 = cos(coord1)cos(coord2)
\item cart2 = cos(coord1)sin(coord2)
\item cart3 = sin(coord1)
}
}
