% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transformations.R
\name{get_spherical}
\alias{get_spherical}
\title{Function to get spherical coordinates of arcs from cartesian coordinates}
\usage{
get_spherical(u)
}
\arguments{
\item{u}{3 column matrix of cartesian coordinates of arcs}
}
\description{
Function to get spherical coordinates of arcs from cartesian coordinates
}
