% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rss.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Function to plot estimated gamma distribution of step lengths}
\usage{
plot_gamma(model, data, obs_data = TRUE)
}
\arguments{
\item{model}{fitted ssf}

\item{data}{data frame with columns: x, y, z, ID}

\item{obs}{whether or not to plot a histogram of the observed data}
}
\description{
Function to plot estimated gamma distribution of step lengths
}
