% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geodesics.R
\name{plot_geodesics}
\alias{plot_geodesics}
\title{Function to plot orthodromic arc (i.e., 3D turning angles) as pts on a sphere}
\usage{
plot_geodesics(data, color = NULL, format = "df")
}
\arguments{
\item{data}{data frame with columns: omega, delta OR a matrix of coordinates
omega is the arc size and delta is the arc orientation}

\item{format}{Format of data (described above). Options are "matrix" or "df"}
}
\description{
Function to plot orthodromic arc (i.e., 3D turning angles) as pts on a sphere
}
