% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rss.R
\name{plot_rss}
\alias{plot_rss}
\title{Function to plot estimated relative selection strength}
\usage{
plot_rss(model, data, variable, log = FALSE)
}
\arguments{
\item{model}{fitted ssf}

\item{data}{data frame with columns: x, y, z, ID}

\item{variable}{which variable to plot}

\item{log}{whether or not to plot the rss on the log scale}
}
\description{
Function to plot estimated relative selection strength
}
