% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for fitting 3D CRWs or SSFs}
\usage{
prep_data(data, n_random = 25, distr = "gamma", par = NULL, ssf = TRUE)
}
\arguments{
\item{data}{data frame of locations with columns: ID, x,y,z, time}

\item{n_random}{Number of random (i.e., integration) points per observation}

\item{distr}{Name(s) of distribution(s) used to generate random points.
Current support for uniform in a sphere ("uniform") or gamma step lengths 
and uniform turning angles ("gamma") or gamma steps and kent arcs ("gamma+kent")}

\item{par}{parameters of the sampling distribution}

\item{ssf}{whether or not to prep the data for an SSF (when set to TRUE, 
random points will be generated; when set to FALSE, movement variables for the 
observed data will be calculated)}
}
\value{
data
}
\description{
Prepare data for fitting 3D CRWs or SSFs
}
