% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_points.R
\name{random_points}
\alias{random_points}
\title{Simulate random locations for model fitting}
\usage{
random_points(obs, n_random, distr = "gamma", par = NULL)
}
\arguments{
\item{obs}{Data frame of observations (ID, x, y, z, time)}

\item{n_random}{Number of random (i.e., integration) points per observation}

\item{distr}{Name(s) of distribution(s) used to generate control steps.
Current support for uniform in a sphere ("uniform") or gamma step lengths 
and uniform turning angles ("gamma") or gamma steps and kent arcs ("gamma+kent")}

\item{par}{parameters of the sampling distribution}
}
\description{
Simulate random locations for model fitting
}
