% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rKent.R
\name{rKent}
\alias{rKent}
\title{Simulate from a Kent distribution}
\usage{
rKent(n, kappa, rho)
}
\arguments{
\item{n}{number of samples}

\item{kappa}{concentration parameter}

\item{rho}{ovalness parameter}
}
\value{
Samples from a Kent distribution (Cartesian coordinates)
}
\description{
Simulate from a Kent distribution
}
\details{
This function assumes mean = (0,0,1), 
major = (1,0,0), and minor = (0,1,0)
}
