% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{setup_samp}
\alias{setup_samp}
\title{Setup sampling of 2D random points}
\usage{
setup_samp(distr = "uniform", obs)
}
\arguments{
\item{distr}{Name of distribution(s) to generate random points}

\item{obs}{Data frame of observed positions, used to get empirical
distributions of movement variables}
}
\value{
List of functions and parameters for sampling
}
\description{
Setup sampling of 2D random points
}
