% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sf_to_df}
\alias{sf_to_df}
\title{Convert SF to data.frame}
\usage{
sf_to_df(data, coords = c("X", "Y"))
}
\arguments{
\item{data}{sf (points only) object}

\item{coords}{vector of X and Y names}
}
\value{
Returns data.frame with all columns, sf geometry dropped, and replaced with coord columns
}
\description{
Brief description of the function.
}
\examples{
# Example usage of the function
myFunction(arg1, arg2)
}
