% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transformations.R
\name{shift_angle}
\alias{shift_angle}
\title{Shift angles to be between -pi and pi}
\usage{
shift_angle(angles)
}
\arguments{
\item{angles}{Vector of angles}
}
\value{
Vector of shifted angles
}
\description{
Shift angles to be between -pi and pi
}
