% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cov.R
\name{sim_cov}
\alias{sim_cov}
\title{Generate a covariate in 3d}
\usage{
sim_cov(nx = 100, ny = 100, nz = 20, window_size = 5)
}
\arguments{
\item{nx}{Number of grid cells in x dimension}

\item{ny}{Number of grid cells in y dimension}

\item{nz}{Number of grid cells in z dimension}

\item{window_size}{Diameter of moving window spherical kernel}
}
\value{
3D array of covariate values, with nx rows, ny columns, and nz layers
}
\description{
Generate a covariate in 3d
}
