% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ssf.R
\name{sim_ssf}
\alias{sim_ssf}
\title{Simulate a movement track from a 3D SSF}
\usage{
sim_ssf(
  xyz0,
  n_zeros,
  n_locs,
  formula,
  betas,
  R = NULL,
  pts_dist = "gamma",
  cov_2D = NULL,
  cov_3D = NULL,
  print = TRUE,
  log = FALSE
)
}
\arguments{
\item{xyz0}{starting location of the track}

\item{n_zeros}{number of proposed endpoints}

\item{n_locs}{number of locations to simulate}

\item{formula}{SSF formula}

\item{betas}{SSF parameters (assumes the order beta_step, beta_logstep, 
beta_omega, beta_omegadelta)}

\item{R}{radius of the sphere (in which to generate endpoints, only 
needed for uniform)}

\item{pts_dist}{Distriution of distances (default = "gamma", also supports 
"uniform")}

\item{cov_2D}{named list of raster stacks of 3D covariate data (optional)}

\item{print}{whether or not to print progress bar}

\item{log}{whether or not to use log in the ssf probs}
}
\value{
Data frame with columns "x", "y", "z", "step", "vbear", 
"hbear", "omega", and "delta"
}
\description{
Simulate a movement track from a 3D SSF
}
