% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{split_at_gap}
\alias{split_at_gap}
\title{Split track at gaps}
\usage{
split_at_gap(data, max_gap = 60, shortest_track = 0)
}
\arguments{
\item{data}{Data frame with (at least) columns for "ID" and "time"}

\item{max_gap}{Longest allowed gap, in minutes (track will be split at longer gaps)}

\item{shortest_track}{Shortest track to keep after splitting, in minutes. Shorter
tracks will be removed from the output data set.}
}
\value{
Data frame with identical structure as input, where ID column
has been replaced by new ID for split tracks. Old ID still accessible as
ID_old column
}
\description{
Split track at gaps
}
