% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{transform_betas}
\alias{transform_betas}
\title{transform betas to the parameters of gamma distribution}
\usage{
transform_betas(beta_step, beta_logstep)
}
\arguments{
\item{beta_step}{coefficient for step}

\item{beta_logstep}{coefficient for log(step)}
}
\description{
transform betas to the parameters of gamma distribution
}
