% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{transform_mean}
\alias{transform_mean}
\title{Transform mean/sd into betas and shape/scale}
\usage{
transform_mean(mean, sd)
}
\arguments{
\item{mean}{mean}

\item{sd}{standard deviation}
}
\value{
beta_step, beta_logstep, shape, and scale
}
\description{
Transform mean/sd into betas and shape/scale
}
