% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{transformCRS}
\alias{transformCRS}
\title{Convert location coordinates to that of environemntal data}
\usage{
transformCRS(data, CRS, newCRS)
}
\arguments{
\item{data}{data frame of locs with x,y columns}

\item{CRS}{current CRS of data}

\item{newCRS}{new CRS}
}
\description{
Convert location coordinates to that of environemntal data
}
