% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var3D.R
\name{var3D}
\alias{var3D}
\title{Functions to calculate 3D variables 
(arc size/orientation, bearings, turning angles)}
\usage{
var3D(data)
}
\arguments{
\item{data}{data frame of locs with ID, x,y,z columns}
}
\description{
Functions to calculate 3D variables 
(arc size/orientation, bearings, turning angles)
}
