% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cov.R
\name{array_to_stack}
\alias{array_to_stack}
\title{Convert 3D array to terra raster stack}
\usage{
array_to_stack(array, xmin, xmax, ymin, ymax, depth)
}
\arguments{
\item{array}{Array of covariate values}

\item{xmin}{Minimum x coordinate}

\item{xmax}{Maximum x coordinate}

\item{ymin}{Minimum y coordinate}

\item{ymax}{Maximum y coordinate}

\item{depth}{Vector of depths (z values)}
}
\value{
SpatRaster stack with one raster for each array layer
}
\description{
Convert 3D array to terra raster stack
}
