% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var3D.R
\name{calc_delta}
\alias{calc_delta}
\title{Function to calculate arc size (omega)}
\usage{
calc_delta(omega, vbear, hbear)
}
\arguments{
\item{omega}{arc size}

\item{vbear}{vertical bearing}

\item{hbear}{horizontal bearing}
}
\description{
Function to calculate arc size (omega)
}
