% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var3D.R
\name{calc_omega}
\alias{calc_omega}
\title{Calculate arc size (omega)}
\usage{
calc_omega(vbear, hbear)
}
\arguments{
\item{vbear}{vertical bearing}

\item{hbear}{horizontal bearing}
}
\value{
Vector of geodesic lengths
}
\description{
Calculate arc size (omega)
}
