% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_df.R
\name{cov_df}
\alias{cov_df}
\title{Create covariate dataframe}
\usage{
cov_df(formula, cov_data, data, dim = 2)
}
\arguments{
\item{formula}{model formula with covariate names}

\item{cov_data}{list of covariate rasters}

\item{data}{location data (either matrix or dataframe with x, y)}

\item{dim}{how many dimensions is the covariate data (2 or 3)}
}
\description{
Create covariate dataframe
}
