% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transformations.R
\name{get_cartesian}
\alias{get_cartesian}
\title{Function to get Cartesian coordinates from spherical coordinates based
on the formulas
cart1 = sin(coord1)cos(coord2),
cart2 = sin(coord1)sin(coord2), 
cart3 = cos(coord1)}
\usage{
get_cartesian(coord1, coord2, type = "geodesics")
}
\arguments{
\item{coord1}{first spherical coordinate}

\item{coord2}{second spherical coordinate}

\item{type}{whether it is bearings or geodesics}
}
\description{
Function to get Cartesian coordinates from spherical coordinates based
on the formulas
cart1 = sin(coord1)cos(coord2),
cart2 = sin(coord1)sin(coord2), 
cart3 = cos(coord1)
}
