% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transformations.R
\name{get_spherical}
\alias{get_spherical}
\title{Get spherical coordinates of arcs from cartesian coordinates}
\usage{
get_spherical(u)
}
\arguments{
\item{u}{3 column matrix of cartesian coordinates of arcs}
}
\value{
Data frame with columns 'omega' (geodesic length) and 'delta'
(geodesic orientation)
}
\description{
Get spherical coordinates of arcs from cartesian coordinates
}
