% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bearings.R
\name{plot_bearings}
\alias{plot_bearings}
\title{Plot 3D bearings as points on a sphere}
\usage{
plot_bearings(data, color = NULL, format = "df")
}
\arguments{
\item{data}{data frame with columns 'vbear' and 'hbear', OR a matrix 
of 3D coordinates}

\item{format}{Format of data (described above). Options are "matrix" or "df"}
}
\description{
Plot 3D bearings as points on a sphere
}
