% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Function to plot 3D location data}
\usage{
plot_data(data, scale = TRUE)
}
\arguments{
\item{data}{data frame with columns: x, y, z, ID}

\item{scale}{TRUE/FALSE to indicate whether x/y/z dimensions should be 
on the same scale (default = TRUE)}
}
\description{
Function to plot 3D location data
}
