% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dKent.R
\name{plot_dKent}
\alias{plot_dKent}
\title{Function to plot density of a Kent distribution on a sphere}
\usage{
plot_dKent(kappa, rho)
}
\arguments{
\item{kappa}{angular concentration parameter}

\item{rho}{ovalness parameter}
}
\value{
plot of the Kent density on a unit sphere
}
\description{
Function to plot density of a Kent distribution on a sphere
}
