% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_3D.R
\name{pts_3D}
\alias{pts_3D}
\title{Function to generate 3D points around a starting coordinate}
\usage{
pts_3D(n, xyz0, dist = "gamma", step_par, angle_par = NULL)
}
\arguments{
\item{n}{number of points to generate}

\item{xyz0}{starting coordinates (x, y, z)}

\item{dist}{distributions of points ("uniform" for spatially uniform points, 
"gamma" for gamma steps and uniform angles, "gamma+kent" for gamma steps and kent angles)}

\item{step_par}{parameters for the step length distribution (needs R for uniform, shape and scale for gamma)}

\item{angle_par}{parameters of the kent distribution (needs kappa, rho, hbear0, and vbear0)}
}
\description{
Function to generate 3D points around a starting coordinate
}
