% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rKent.R
\name{rKent}
\alias{rKent}
\title{Function to simulate samples from a kent distribution 
(assumes mean = (0,0,1), major = (1,0,0), minor = (0,1,0))}
\usage{
rKent(n, kappa, rho)
}
\arguments{
\item{n}{number of samples}

\item{kappa}{concentration parameter}

\item{rho}{ovalness parameter}
}
\description{
Function to simulate samples from a kent distribution 
(assumes mean = (0,0,1), major = (1,0,0), minor = (0,1,0))
}
