% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_crw.R
\name{sim_crw}
\alias{sim_crw}
\title{Function to simulate a movement track from a 3D correlated random walk}
\usage{
sim_crw(
  xyz0 = c(0, 0, 0),
  n_locs,
  type,
  step_par,
  kappa = NULL,
  rho = NULL,
  kappa2 = NULL
)
}
\arguments{
\item{xyz0}{starting location of the track (default = c(0, 0, 0))}

\item{n_locs}{number of locations to simulate}

\item{type}{type of correlated random walk (options: "rw", "vMF", "kent", "bcrw")}

\item{step_par}{parameters of step length distribution in format c(shape, scale)}

\item{kappa}{concentration parameter}
}
\description{
Function to simulate a movement track from a 3D correlated random walk
}
