% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{transform_par}
\alias{transform_par}
\title{Transform shape and scale into betas and mean/sd}
\usage{
transform_par(shape, scale)
}
\arguments{
\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\value{
beta_step, beta_logstep, mean, and sd
}
\description{
Transform shape and scale into betas and mean/sd
}
