% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{transformCRS}
\alias{transformCRS}
\title{Convert 2D spatial locations to new CRS}
\usage{
transformCRS(data, CRS, newCRS)
}
\arguments{
\item{data}{data frame of locs with x,y columns}

\item{CRS}{current CRS of data}

\item{newCRS}{new CRS}
}
\value{
Data frame where x and y have been transformed to new CRS
}
\description{
Convert 2D spatial locations to new CRS
}
